✨ feat: implements atomic architeture, removes inline styles && duplicate code
