+++ src/pages/relatorios.tsx
@@ -1,4 +1,4 @@
-import Header from "@/app/components/Header";
+import Header from "@/app/organisms/Header";
 
 export default function Relatorios() {
   return (
