+++ tailwind.config.ts
@@ -4,8 +4,10 @@ import type { Config } from "tailwindcss";
 
 const config: Config = {
   content: [
-    "./src/app/**/*.{js,ts,jsx,tsx,mdx}", // This will scan the app directory
-    "./src/app/components/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/app/atoms/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/app/molecules/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/app/organisms/**/*.{js,ts,jsx,tsx,mdx}",
     flowbite.content(),
   ],
   theme: {
@@ -15,6 +17,15 @@ const config: Config = {
       },
     },
   },
+  safelist: [
+    "px-8",
+    "sm:px-16",
+    "md:px-24",
+    "lg:px-32",
+    "xl:px-64",
+    "pl-7",
+    "pt-8",
+  ],
   plugins: [flowbite.plugin()],
 };
 export default config;
